Mapeando Produto

+++ DOMAIN/Models/Produto.cs
@@ -5,8 +5,7 @@ using System.Collections.Generic;
 namespace Base.DOMAIN.Models
 {
     public class Produto
-    {
-        
+    { 
         public int IdProduto { get; set; }
         public string NomeProduto { get; set; }
         public double ValorProduto { get; set; }
@@ -14,7 +13,7 @@ namespace Base.DOMAIN.Models
         public int QuantidadeVendas { get; set; }
         public string DescricaoProduto { get; set; }
         public int IdFabricante { get; set; }
-        public int Avaliacao {  get; set; }
+        public byte Avaliacao {  get; set; }
         public byte[] FotoProduto { get; set; }
         public CategoriaProduto CategoriaProduto { get; set; }
         public virtual ICollection<ItemVenda> VendaProduto { get; set; }
